# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Templarc",
...
    "contact": {
      "name": "Templarc"
    },
    "license": {
      "name": "Private"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/catalog/projects": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List projects",
        "description": "Return all projects. Optionally filter by organization and/or a name search string.",
        "operationId": "list_projects_catalog_projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by organization",
              "title": "Organization Id"
            },
            "description": "Filter by organization"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Partial match on project name",
              "title": "Search"
            },
            "description": "Partial match on project name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response List Projects Catalog Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Create project",
        "description": "Create a new project. Initialises a Git subdirectory at `git_path` (defaults to `name`) with a .gitkeep commit so the directory exists before any templates are written.",
        "operationId": "create_project_catalog_projects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/projects/{project_id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get project",
        "description": "Fetch a project by ID, including its full template hierarchy tree.",
        "operationId": "get_project_catalog_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Catalog"
        ],
        "summary": "Update project",
        "description": "Partial update of project metadata. `name` and `organization_id` are immutable after creation.",
        "operationId": "update_project_catalog_projects__project_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/projects/{project_id}/templates": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List templates as tree",
        "description": "Return the template hierarchy for a project as a nested tree. Only active templates are included. Nodes are ordered by sort_order then name within each level.",
        "operationId": "get_project_templates_catalog_projects__project_id__templates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateTreeNode"
                  },
                  "title": "Response Get Project Templates Catalog Projects  Project Id  Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/{project_slug}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get product catalog",
        "description": "Return the full product catalog for a project, identified by its slug (`name` field). Intended for end-user selection UIs.\n\nEach template entry includes:\n- `breadcrumb` \u2014 ordered list of display names from root ancestor to this node\n- `parameter_count` \u2014 active template-scope parameters registered\n- `has_remote_datasources` \u2014 whether the `.j2` frontmatter declares `data_sources`\n- `is_leaf` \u2014 `true` when the template has no active child templates\n\nBoth leaf templates (selectable by end users) and intermediate nodes (for hierarchical display) are included.",
        "operationId": "get_product_catalog_catalog__project_slug__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogProjectOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "display_name",
          "description"
        ],
        "title": "CatalogProjectOut",
        "description": "Minimal project summary embedded in the catalog response."
      },
      "CatalogResponse": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/CatalogProjectOut"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/CatalogTemplateItem"
            },
            "type": "array",
            "title": "Templates"
          }
        },
        "type": "object",
        "required": [
          "project",
          "templates"
        ],
        "title": "CatalogResponse",
        "description": "Full product catalog response for a single project."
      },
      "CatalogTemplateItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "breadcrumb": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Breadcrumb",
            "description": "Ordered list of display_names from root to this template (inclusive)."
          },
          "parameter_count": {
            "type": "integer",
            "title": "Parameter Count",
            "description": "Count of active template-scope parameters registered for this template.",
            "default": 0
          },
          "has_remote_datasources": {
            "type": "boolean",
            "title": "Has Remote Datasources",
            "description": "True if the template's .j2 frontmatter declares one or more data_sources.",
            "default": false
          },
          "is_leaf": {
            "type": "boolean",
            "title": "Is Leaf",
            "description": "True when this template has no active child templates.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "display_name",
          "description"
        ],
        "title": "CatalogTemplateItem",
        "description": "One template entry in the product catalog.\n\nIncludes enough information for a selection UI without requiring a\nsecond API call:\n  - breadcrumb: display_names from root ancestor down to this node\n  - parameter_count: number of active template-scope parameters registered\n  - has_remote_datasources: True if the .j2 frontmatter declares data_sources\n  - is_leaf: True when the template has no active children"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name",
            "description": "Slug-style identifier, unique per org"
          },
          "display_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "git_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Path",
            "description": "Subdirectory inside the templates repo. Defaults to `name` if omitted."
          },
          "output_comment_style": {
            "type": "string",
            "maxLength": 10,
            "title": "Output Comment Style",
            "description": "Comment prefix used in rendered output headers: #, !, //, or none",
            "default": "#"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "name",
          "display_name"
        ],
        "title": "ProjectCreate",
        "examples": [
          {
            "description": "Cisco IOS router configuration templates",
            "display_name": "Router Provisioning",
            "git_path": "router_provisioning",
            "name": "router_provisioning",
            "organization_id": 1,
            "output_comment_style": "!"
          }
        ]
      },
      "ProjectDetailOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "git_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Path"
          },
          "output_comment_style": {
            "type": "string",
            "title": "Output Comment Style"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/TemplateTreeNode"
            },
            "type": "array",
            "title": "Templates",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "name",
          "display_name",
          "description",
          "git_path",
          "output_comment_style",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectDetailOut",
        "description": "Project with its full template hierarchy tree.",
        "examples": [
          {
            "created_at": "2024-01-15T09:00:00Z",
            "description": "Cisco IOS router configuration templates",
            "display_name": "Router Provisioning",
            "git_path": "router_provisioning",
            "id": 1,
            "name": "router_provisioning",
            "organization_id": 1,
            "output_comment_style": "!",
            "updated_at": "2024-01-15T09:00:00Z"
          }
        ]
      },
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "git_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Path"
          },
          "output_comment_style": {
            "type": "string",
            "title": "Output Comment Style"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "name",
          "display_name",
          "description",
          "git_path",
          "output_comment_style",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut",
        "examples": [
          {
            "created_at": "2024-01-15T09:00:00Z",
            "description": "Cisco IOS router configuration templates",
            "display_name": "Router Provisioning",
            "git_path": "router_provisioning",
            "id": 1,
            "name": "router_provisioning",
            "organization_id": 1,
            "output_comment_style": "!",
            "updated_at": "2024-01-15T09:00:00Z"
          }
        ]
      },
      "ProjectUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "git_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Path"
          },
          "output_comment_style": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Comment Style"
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
...
      "TemplateTreeNode": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/TemplateTreeNode"
            },
            "type": "array",
            "title": "Children",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "display_name",
          "is_active",
          "sort_order"
        ],
        "title": "TemplateTreeNode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}